Implemented Inventory Database
Code is almost identical to the username/password database. Using a
different .ser file, namely "Inventory.ser" . Application inserts the
hardcoded inventory into the database when the "mainInventory" Inventory
object gets initialized, as part of its default constructor. We will be
able to reuse this same code, or something extremely similar, with the
action listeners where a seller adds a new product to the databse during
runtime.