Lots of inventory work
Added ProductList interface and ConcreteList abstract class that
implements it, and changed Inventory to extend ConcreteList. Also moved
the code for saving to/loading from DB (that is, serializing and
un-serializing) into "saveToDB" and "loadDB" methods inside Inventory,
which can be called whenever a load or save needs to happen. On that
subject, also changed this code so that the prodList itself is what gets
serialized (which apparently automatically serializes everything in it,
too!), rather than iterating through the prodList and serializing all
the products in the list individually. Finally, CustomerInventoryView
now creates an instance of Inventory and adds JLabels for each product
in the inventory.