CheckOutView coming along
Looks more like it should. Lists the items along with their quantity in
a JLabel. Added a remove button but it has no functionality yet. Added a
submit order button but it has no functionality yet. I also need to
eventually consolidate all of this code into the member functions of the
ShoppingCart class.