Added "SaveToDB" functionality for checkout
I realized that the app needs to keep a permanent record of all sales in
order to get the "revenue" figure for the sales data view. So I added a
saveToDB function in ShoppingCart class that gets called as part of the
overall checkout function, and a function to get the total sell price of
all sales for the sales data view.