Javadoc comments and some updates:
Added Javadoc comments throughout.

Modified createAccount method so that an account with username "seller"
is automatically created as a seller-type account.

Modified Inventory's saveToDB method so that the first time - and only
the first time - that the app runs, it creates and saves a few products
to initially populate the inventory.

Changed filepaths on all serialize-related methods to automatically save
to the the user's "User" folder.